Fix tests in FindICONV.cmake when cross-compiling
Fix the tests that are performed when CMake is in cross-compiling mode.
Since much of FindICONV.cmake is ported from Gnulib (which uses Autoconf),
the "host" platform (the value of the '--host=' option argument to an
Autoconf-generated 'configure' script) was incorrectly presumed to be
equivalent to what CMake calls the "host" platform (resulting in the use of
the CMAKE_HOST_SYSTEM variable instead of CMAKE_SYSTEM). In reality, the
GNU Build System's "build" platform more closely corresponds to CMake's
"host", while GNU's "host" corresponds to CMake's "target".